# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SessionResponse": {
        "description": "The current session. `token` is present only on the response that created it.",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "issued_via": {
            "title": "Issued Via",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "display_name",
          "roles",
          "expires_at",
          "issued_via"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "StubSignInRequest": {
        "additionalProperties": false,
        "description": "Who to sign in. There is no password field, and that is `\u00a712.2` rather than an omission.",
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "StubSignInRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Application-owned sales workflow. Shadow mode; live outreach is gated.",
    "title": "Matrix Power Always-On AI Sales Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/session": {
      "delete": {
        "description": "Revoke the caller's own session, and only their own.\n\nA caller with no session gets `204` rather than `401`: signing out when already signed out is\nthe state they asked for, and answering `401` would make a dashboard sign-out button fail for\na reviewer whose session had just expired.\n\nBoth cookies are cleared **before** the `principal is None` return, so a caller holding a\ncookie the server no longer honours still gets rid of it. Leaving a dead session cookie in the\nbrowser is how a reviewer ends up signed out and unable to sign in again without knowing to\nclear cookies by hand.",
        "operationId": "sign_out_api_auth_session_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign out, revoking the current session",
        "tags": [
          "auth"
        ]
      },
      "get": {
        "description": "Who the caller is. `401` when nobody.\n\nNot `requires(...)`: there is no permission to hold here \u2014 the question is whether a session\nresolves at all, and every role may ask it. The dashboard uses this to tell \"signed out\" from\n\"signed in without the role\", which are different screens.",
        "operationId": "read_session_api_auth_session_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The current session, or 401",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/stub-sign-in": {
      "post": {
        "description": "Issue a session for a known local user. Verifies nothing; refused outside `local`.\n\nThe environment check is the first statement on purpose \u2014 see the module docstring. An unknown\nemail is `404` and creates nobody: the roster is `Q-005` and `Q-026`, not whoever can reach\nthis port.",
        "operationId": "stub_sign_in_endpoint_api_auth_stub_sign_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StubSignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign in as an existing local user (local environment only)",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}